refactor: move AsyncVFSRouter from services/routing/ to core/
AsyncVFSRouter is a thin async adapter for the kernel-tier PathRouter,
conforming to VFSRouterProtocol (a kernel interface per KERNEL-ARCHITECTURE.md §3).
It has zero service-layer logic and belongs in nexus.core, not nexus.services.

- Move services/routing/async_router.py → core/async_router.py
- Delete empty services/routing/ package
- Rewrite all imports (factory, unit tests, e2e tests, mock paths)
